OPENNLP-1888: Place annotators with their components
