+++ opennlp-core/opennlp-runtime/src/main/java/opennlp/tools/tokenize/TokenizerAnnotator.java
@@ -15,13 +15,17 @@
  * limitations under the License.
  */
 
-package opennlp.tools.document;
+package opennlp.tools.tokenize;
 
 import java.util.ArrayList;
 import java.util.List;
 import java.util.Set;
 
-import opennlp.tools.tokenize.Tokenizer;
+import opennlp.tools.document.Annotation;
+import opennlp.tools.document.Document;
+import opennlp.tools.document.DocumentAnnotator;
+import opennlp.tools.document.LayerKey;
+import opennlp.tools.document.Layers;
 import opennlp.tools.util.Span;
 
 /**
